+++ examples/variables.ny
@@ -1,3 +1,9 @@
+ndondomeko phatikiza(yoyamba, chiwiri) {
+	bweza yoyamba + chiwiri;
+}
+
 nambala x = 5;
 nambala y = 10;
 mawu foobar = "Iyi ndi foo bar";
+
+phatikiza(x, y);
