+++ src/ast/function.go
@@ -1,0 +1,31 @@
+package ast
+
+import (
+	"bytes"
+	"github.com/sevenreup/chewa/src/token"
+	"strings"
+)
+
+type FunctionLiteral struct {
+	Token      token.Token // The 'ndondomeko' token
+	Name       *Identifier // The name of the function
+	Parameters []*Identifier
+	Body       *BlockStatement
+}
+
+func (fl *FunctionLiteral) expressionNode()      {}
+func (fl *FunctionLiteral) TokenLiteral() string { return fl.Token.Literal }
+func (fl *FunctionLiteral) String() string {
+	var out bytes.Buffer
+	params := []string{}
+	for _, p := range fl.Parameters {
+		params = append(params, p.String())
+	}
+	out.WriteString(fl.TokenLiteral())
+	out.WriteString(fl.Name.TokenLiteral())
+	out.WriteString("(")
+	out.WriteString(strings.Join(params, ", "))
+	out.WriteString(") ")
+	out.WriteString(fl.Body.String())
+	return out.String()
+}
