+++ src/ast/integer.go
@@ -1,0 +1,12 @@
+package ast
+
+import "github.com/sevenreup/chewa/src/token"
+
+type IntegerLiteral struct {
+	Token token.Token
+	Value int64
+}
+
+func (il *IntegerLiteral) expressionNode()      {}
+func (il *IntegerLiteral) TokenLiteral() string { return il.Token.Literal }
+func (il *IntegerLiteral) String() string       { return il.Token.Literal }
