+++ src/ast/string.go
@@ -1,0 +1,12 @@
+package ast
+
+import "github.com/sevenreup/chewa/src/token"
+
+type StringLiteral struct {
+	Token token.Token
+	Value string
+}
+
+func (sl *StringLiteral) expressionNode()      {}
+func (sl *StringLiteral) TokenLiteral() string { return sl.Token.Literal }
+func (sl *StringLiteral) String() string       { return sl.Token.Literal }
