+++ src/main.go
@@ -2,6 +2,9 @@ package main
 
 import (
 	"fmt"
+	"github.com/sevenreup/chewa/src/evaluator"
+	"github.com/sevenreup/chewa/src/object"
+	"github.com/sevenreup/chewa/src/utils"
 	"log"
 	"os"
 
@@ -16,8 +19,14 @@ func main() {
 	}
 	l := lexer.New(file)
 	p := parser.New(l)
+	env := object.NewEnvironment()
 	program := p.ParseProgram()
-	for _, v := range program.Statements {
-		fmt.Printf("%+v\n", v)
+	if len(p.Errors()) != 0 {
+		utils.PrintParserErrors(os.Stdout, p.Errors())
+	}
+	evaluated := evaluator.Eval(program, env)
+	if evaluated != nil {
+		fmt.Print(evaluated.Inspect())
+		fmt.Print("\n")
 	}
 }
